pyproject.toml
[tool.poetry]
name = "celerygeneratecsv"
version = "0.1.0"
description = ""
authors = ["Anakion <andreygolosov786@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
flower = "^2.0.1"
python-dotenv = "^1.0.1"
redis = "^5.2.1"
celery = "^5.4.0"
uvicorn = "^0.32.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
